requirements.txt
streamlit
groq
tavily-python
duckduckgo-search